[types] Refactor AccountAddress to move_core_types
