+++ language/move-core/types/Cargo.toml
@@ -11,6 +11,9 @@ edition = "2018"
 
 [dependencies]
 anyhow = "1.0"
+bytes = "0.5.4"
+hex = "0.4.2"
+rand = "0.7.3"
 proptest = { version = "0.9.6", default-features = false, optional = true }
 mirai-annotations = "1.8.0"
 proptest-derive = { version = "0.1.2", default-features = false, optional = true }
@@ -19,6 +22,8 @@ serde = { version = "1.0.106", default-features = false }
 
 lcs = { path = "../../../common/lcs", version = "0.1.0", package = "libra-canonical-serialization" }
 libra-workspace-hack = { path = "../../../common/workspace-hack", version = "0.1.0" }
+libra-crypto = { path = "../../../crypto/crypto", version = "0.1.0" }
+libra-crypto-derive = { path = "../../../crypto/crypto-derive", version = "0.1.0" }
 
 [dev-dependencies]
 once_cell = "1.3.1"
