+++ language/move-core/types/src/lib.rs
@@ -3,6 +3,7 @@
 
 //! Core types for Move.
 
+pub mod account_address;
 pub mod gas_schedule;
 pub mod identifier;
 #[cfg(test)]
