+++ language/move-core/types/src/unit_tests/mod.rs
@@ -1,4 +1,5 @@
 // Copyright (c) The Libra Core Contributors
 // SPDX-License-Identifier: Apache-2.0
 
+mod address_test;
 mod identifier_test;
