+++ testsuite/cluster-test/src/tx_emitter.rs
@@ -573,7 +573,7 @@ fn gen_mint_txn_request(
 fn gen_random_account(rng: &mut StdRng) -> AccountData {
     let key_pair = KeyPair::generate(rng);
     AccountData {
-        address: AccountAddress::from_public_key(&key_pair.public_key),
+        address: libra_types::account_address::from_public_key(&key_pair.public_key),
         key_pair,
         sequence_number: 0,
     }
